+++ src/view/user/setting/StudyList/StudyListTable/StudyListTable.tsx
