+++ src/view/user/setting/StudyList/StudyListTable/constant.tsx
@@ -6,10 +6,10 @@ import {
 import { pinStyle } from "@/shared/component/Table/index.css";
 import type { StudyListType, TableDataType } from "@/shared/type/table";
 import { format } from "date-fns";
-import SortIcon from "../SortIcon";
-import StatusDropdonwMenu from "../StatusDropdownMenu";
-import { textStyle } from "../StatusDropdownMenu/index.css";
-import StatusIcon from "../StatusIcon";
+import SortIcon from "../../SortIcon";
+import StatusDropdownMenu from "../../StatusDropdownMenu";
+import { textStyle } from "../../StatusDropdownMenu/index.css";
+import StatusIcon from "../../StatusIcon";
 import { useTableDispatch, useTableState } from "./hook";
 
 export const STUDY_LIST_COLUMNS: TableDataType<StudyListType>[] = [
@@ -110,7 +110,7 @@ export const STUDY_LIST_COLUMNS: TableDataType<StudyListType>[] = [
   },
   {
     key: "status",
-    Header: () => <StatusDropdonwMenu />,
+    Header: () => <StatusDropdownMenu />,
     Cell: (data) => <StatusIcon status={data.status} />,
     width: 100,
   },
